fillout predictors list view and give it a route
