+++ neuroscout/frontend/src/browser/DatasetDetailView.tsx
@@ -5,7 +5,8 @@ import { Collapse, Row, Table, Tag } from 'antd'
 import { AppAnalysis, Dataset, Predictor, Task } from '../coretypes'
 import { api } from '../api'
 import { predictorColor } from '../utils'
-import { AnalysisListTable, PredictorTagList } from './AnalysisList'
+import { AnalysisListTable } from './AnalysisList'
+import { PredictorByExtractorList, PredictorTagList } from './PredictorList'
 
 const { Panel } = Collapse
 
@@ -24,32 +25,6 @@ export interface Dataset {
   percentFemale?: number
 }
 */
-
-export const PredictorList = (props: {
-  predictors: Predictor[]
-}): JSX.Element => {
-  const extractors = {}
-  props.predictors.map(predictor => {
-    let ef_name = 'none'
-    if (predictor.extracted_feature?.extractor_name) {
-      ef_name = predictor.extracted_feature.extractor_name
-    } else if (predictor.source) {
-      ef_name = predictor.source
-    }
-    extractors[ef_name]
-      ? extractors[ef_name].push(predictor)
-      : (extractors[ef_name] = [predictor])
-  })
-  return (
-    <Collapse style={{ width: '100%' }} ghost>
-      {Object.keys(extractors).map(key => (
-        <Panel key={key} header={key}>
-          <PredictorTagList predictors={extractors[key]} />
-        </Panel>
-      ))}
-    </Collapse>
-  )
-}
 
 interface TaskPredictors {
   task: Task
@@ -105,7 +80,7 @@ export const DatasetDetailView = (props: Dataset): JSX.Element => {
           <Row key={taskPredictor.task.id}>
             <div className="viewTitle">{taskPredictor.task.name}</div>
             <br />
-            <PredictorList predictors={taskPredictor.predictors} />
+            <PredictorByExtractorList predictors={taskPredictor.predictors} />
           </Row>
         ))}
       </MainCol>
